+++ go/api/glide_client.go
@@ -10,22 +10,12 @@ import (
 	"github.com/valkey-io/valkey-glide/go/api/constants"
 	"github.com/valkey-io/valkey-glide/go/api/models"
 	"github.com/valkey-io/valkey-glide/go/api/options"
+	"github.com/valkey-io/valkey-glide/go/internal/interfaces"
 	"github.com/valkey-io/valkey-glide/go/internal/utils"
 )
 
 // GlideClient interface compliance check.
-var _ GlideClientCommands = (*GlideClient)(nil)
-
-// All commands that can be executed by GlideClient.
-type GlideClientCommands interface {
-	BaseClientCommands
-	GenericCommands
-	ServerManagementCommands
-	BitmapCommands
-	ConnectionManagementCommands
-	ScriptingAndFunctionStandaloneCommands
-	PubSubStandaloneCommands
-}
+var _ interfaces.GlideClientCommands = (*GlideClient)(nil)
 
 // Client used for connection to standalone servers.
 // Use [NewGlideClient] to request a client.